Added new exception to "dataClass" called CritialCourseDataMissingException, which is thrown by DatabaseReader in its findCourse method if no "skema" data could be extracted from the database.
Added the "year"-field to Course class which is either "e","f" or "b". ("e" for "efterår", "f" for "forår" and "b" for "begge") It specifies the time of year the course is available in the database. Blank entries in the database are treated as "b".

DatabaseReader's findCourse now properly finds and fills a Course object with the data from the databases. NB: findCourse may leave null fields in Course.