Commented the Exceptions
Created a SelectedCourse class (extends Course). It contains the semester the course is taken on. It is also comparable/sortable and the order is determined primarily by the semester (the lower semester, the closer to the first entry) and then on tie by the ID number of the course (the lower ID, the lower entry number)