DatabaseReader now implements Iterable and includes an inner class iter, which (while not following the Iterable interface nor its standards) takes care of looking through the Database for the next entries in it. Invalid courses (that would cause DatabaseReader.findCourse(String) to throw an exception) are ignored by this class.
DatabaseReaderIterator is a new class that (as the name implies) iteratorates through all the entries in the Database and fetches all the valid courses in it one by one. Invalid courses (that would cause DatabaseReader.findCourse(String) to throw an exception) are ignored by this Iterator.

To get a DatabaseReaderIterator, use the DatabaseReader.iterator() method.