Added new Exception for when adding a course to a StudyPlan and another course on the same semester have conflicting skema-data.
DatabaseReaderIterator class was removed. DatabaseReader's inner class has replaced it.

UserDatabase, altered some comments.

Core (and ProgramCore), added comments and altered/added some methods. Removed the Ilterator. The effects of it can be achieved via the getAllCourses[] method.

Dialog, removed a catch for a Exception that was no longer thrown by the Core.